Refactor database schema management and improve AdEvent processing: - Updated aggregated_campaign_stats table definition to include primary key and unique constraints. - Enhanced database initialization logic in AdEventSparkStreamer to check and create the table if it doesn't exist. - Improved error handling and logging during schema checks and view management. - Adjusted AdEvent model to implement Serializable and updated field types for better compatibility.

+++ root/q1_realtime_stream_processing/src/main/java/com/tabularasa/bi/q1_realtime_stream_processing/repository/AggregatedCampaignStatsRepository.java
@@ -14,7 +14,7 @@ import java.util.Optional;
  * Repository interface for AggregatedCampaignStats entities.
  */
 @Repository
-public interface AggregatedCampaignStatsRepository extends JpaRepository<AggregatedCampaignStats, Long> {
+public interface AggregatedCampaignStatsRepository extends JpaRepository<AggregatedCampaignStats, Integer> {
 
     /**
      * Find statistics for a specific campaign and event type within a time window.
